% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_halo.R
\name{generate_cell_uuid}
\alias{generate_cell_uuid}
\title{Generate unique cell UUIDs}
\usage{
generate_cell_uuid(dat, cols_uuid)
}
\arguments{
\item{dat}{Dataframe containing the columns.}

\item{cols_uuid}{Column names to use for UUID generation.}
}
\value{
Character vector of UUIDs.
}
\description{
Creates SHA1 hash from specified columns to uniquely identify cells.
}
