% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_manifest.R
\name{plot_cell_counts}
\alias{plot_cell_counts}
\title{Barplot of cell counts per sample}
\usage{
plot_cell_counts(cell_summary)
}
\arguments{
\item{cell_summary}{Tibble from \code{\link[=summarize_cells]{summarize_cells()}}.}
}
\value{
A ggplot object.
}
\description{
Barplot of cell counts per sample
}
