% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_manifest.R
\name{plot_marker_presence}
\alias{plot_marker_presence}
\title{Tile plot of which markers each sample's panel contains}
\usage{
plot_marker_presence(presence_matrix)
}
\arguments{
\item{presence_matrix}{Wide tibble from \code{\link[=marker_presence_matrix]{marker_presence_matrix()}}.}
}
\value{
A ggplot object.
}
\description{
Tile plot of which markers each sample's panel contains
}
