% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annotation.R
\name{plot_tumor_states}
\alias{plot_tumor_states}
\title{Barplot of tumor-cell state flags (the four "respectively" markers)}
\usage{
plot_tumor_states(state_summary)
}
\arguments{
\item{state_summary}{Long tibble from \code{\link[=summarize_states]{summarize_states()}}.}
}
\value{
A ggplot object (or \code{NULL} if there are no tumor states to show).
}
\description{
Focused view of the tumor sub-question: of the tumor cells in each sample,
how many are positive for Ki-67 / pSTAT1 / pSTAT3 / GZMB, respectively.
}
