% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_halo.R
\name{read_halo}
\alias{read_halo}
\title{Read a Halo CSV file}
\usage{
read_halo(file, ...)
}
\arguments{
\item{file}{Path to CSV file.}

\item{...}{Additional arguments passed to readr::read_csv().}
}
\value{
A tibble with cleaned column names.
}
\description{
Reads a Halo-exported CSV and cleans column names.
}
