% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_halo.R
\name{load_halo}
\alias{load_halo}
\title{Load a Halo Object CSV file}
\usage{
load_halo(hfile, uuidCols, sampleName, colsExtra, markerMap)
}
\arguments{
\item{hfile}{Halo CSV file}

\item{uuidCols}{Columns to use to compute cell UUID}

\item{sampleName}{If string set SID, if function SID=sampleName(hfile)}

\item{colsExtra}{Any extra columns to include}

\item{markerMap}{Names vector to rename markers}
}
\value{
A list with elements: cell.data, marker.data, and VERSION
}
\description{
Load a Halo Object CSV file
}
